requirements.txt
PyQt6>=6.5.0
numpy>=1.24.0
scipy>=1.10.0
torch>=2.0.0
torchvision>=0.15.0
PyYAML>=6.0
pylablib>=1.4.1
opencv-python>=4.8.0
matplotlib>=3.7.0
pyqtgraph>=0.13.0
